requirements.txt
appdirs==1.4.4
asttokens==3.0.1
attrs==25.4.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.14.3
beniget==0.4.1
biopython==1.86
bleach==6.3.0
blinker==1.4
Bottleneck==1.3.2
Brotli==1.0.9
certifi==2025.11.12
chardet==4.0.0
charset-normalizer==3.4.4
command-not-found==0.3
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
decorator==5.2.1
defusedxml==0.7.1
distro==1.7.0
distro-info===1.1build1
docopt==0.6.2
et-xmlfile==1.0.1
executing==2.2.1
fair-esm==2.0.0
fastjsonschema==2.21.2
filelock==3.19.1
fonttools==4.29.1
fs==2.4.12
fsspec==2025.9.0
gast==0.5.2
hf-xet==1.2.0
html5lib==1.1
httplib2==0.20.2
huggingface-hub==0.36.0
idna==3.11
importlib-metadata==4.6.4
iniconfig==1.1.1
ipython==8.12.3
jdcal==1.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
kahip==3.22
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llvmlite==0.38.0
lxml==4.8.0
lz4==3.1.3+dfsg
MarkupSafe==2.1.5
matplotlib==3.5.1
matplotlib-inline==0.2.1
mistune==3.1.4
more-itertools==8.10.0
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
netifaces==0.11.0
networkx==3.3
numba==0.55.1
numexpr==2.8.1
numpy==1.26.4
oauthlib==3.2.0
odfpy==1.4.2
olefile==0.46
openpyxl==3.0.9
packaging==25.0
pandas==1.3.5
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.5.0
platformdirs==4.5.1
pluggy==0.13.0
ply==3.11
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.10.0
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.5
python-apt==2.4.0+ubuntu1
python-dateutil==2.9.0.post0
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
pyzmq==27.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scipy==1.8.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
systemd-python==234
tables==3.7.0
tinycss2==1.4.0
tokenizers==0.22.2
toml==0.10.2
tomli==1.2.2
torch==2.2.2+cpu
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.6
typing_extensions==4.15.0
ubuntu-advantage-tools==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==2.6.0
wadllib==1.3.6
wcwidth==0.2.14
webencodings==0.5.1
xlwt==1.3.0
yarg==0.1.9
zipp==1.0.0
